feat: implement repository and services

+++ src/main/java/dev/gymService/dao/implmentations/TraineeRepositoryImpl.java
@@ -2,7 +2,9 @@ package dev.gymService.dao.implmentations;
 
 import dev.gymService.dao.interfaces.TraineeRepository;
 import dev.gymService.model.Trainee;
+import dev.gymService.model.Trainer;
 import dev.gymService.model.Training;
+import dev.gymService.utills.FileLogger;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
@@ -10,10 +12,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Repository;
 
+import java.time.LocalDate;
 import java.util.List;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 @Repository
 public class TraineeRepositoryImpl implements TraineeRepository {
+    private static final Logger logger = FileLogger.getLogger(TraineeRepositoryImpl.class);
     @Autowired
     @Qualifier("getSessionFactory")
     SessionFactory sessionFactory;
@@ -29,43 +35,71 @@ public class TraineeRepositoryImpl implements TraineeRepository {
     }
 
     @Override
-    public Trainee getTraineeById(Long id) {
+    public Trainee getTraineeById(Long id, String userName, String password) {
         try (Session session = sessionFactory.openSession()) {
-            return session.get(Trainee.class, id);
+            //  Check userName and password matching
+            Trainee trainee = session.get(Trainee.class, id);
+            if (trainee.getUserName().equals(userName) && trainee.getPassword().equals(password)) {
+                logger.log(Level.INFO, "Successfull authentification for trainer: " + trainee.getUserName());
+                return trainee;
+            } else {
+                logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
+                return null;
+            }
         }
     }
 
     @Override
-    public Trainee getTraineeByUserName(String userName) {
+    public Trainee getTraineeByUserName(String userName, String password) {
         try (Session session = sessionFactory.openSession()) {
-            return session.createQuery( "SELECT t FROM Trainee t WHERE t.user.userName = :userName", Trainee.class)
-                    .setParameter("userName", userName)
-                    .uniqueResult();
+
+            //  Check userName and password matching
+            Trainee trainee = this.getTraineeByUserName(userName, password);
+            if (trainee.getUserName().equals(userName) && trainee.getPassword().equals(password)) {
+                logger.log(Level.INFO, "Successfull authentification for trainer: " + trainee.getUserName());
+                return trainee;
+            } else {
+                logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
+                return null;
+            }
         }
     }
 
     @Override
-    public void changeTraineePassword(String userName,String newPassword) {
+    public void changeTraineePassword(String userName, String oldPassword, String newPassword) {
         try (Session session = sessionFactory.openSession()) {
             Transaction transaction = session.beginTransaction();
 
-            session.createQuery( "UPDATE User u SET u.password = :newPassword WHERE u.userName = :userName")
-                    .setParameter("newPassword", newPassword)
-                    .setParameter("userName", userName)
-                    .executeUpdate();
-
+            //  Check userName and password matching
+            Trainee trainee = this.getTraineeByUserName(userName, oldPassword);
+            if (trainee.getUserName().equals(userName) && trainee.getPassword().equals(oldPassword)) {
+                logger.log(Level.INFO, "Successfull authentification for trainer: " + trainee.getUserName());
+                session.createQuery("UPDATE User u SET u.password = :newPassword WHERE u.userName = :userName")
+                        .setParameter("newPassword", newPassword)
+                        .setParameter("userName", userName)
+                        .executeUpdate();
+            } else {
+                logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
+            }
             transaction.commit();
         }
     }
 
     @Override
-    public void deleteTraineeByUserName(String userName) {
+    public void deleteTraineeByUserName(String userName, String password) {
         Transaction transaction = null;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
-            session.createQuery("DELETE FROM Trainee WHERE userName = :userName")
-                    .setParameter("userName", userName)
-                    .executeUpdate();
+            //  Check userName and password matching
+            Trainee trainee = this.getTraineeByUserName(userName, password);
+            if (trainee.getUserName().equals(userName) && trainee.getPassword().equals(password)) {
+                session.createQuery("DELETE FROM Trainee WHERE userName = :userName")
+                        .setParameter("userName", userName)
+                        .executeUpdate();
+                logger.log(Level.INFO, "Successfull authentification for trainer: " + trainee.getUserName());
+            } else {
+                logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
+            }
             transaction.commit();
         } catch (Exception e) {
             if (transaction != null) {
@@ -76,16 +110,14 @@ public class TraineeRepositoryImpl implements TraineeRepository {
     }
 
     @Override
-    public void changeTraineeStatus(String userName) {
+    public void changeTraineeStatus(String userName, String password) {
         // Get trainee by userName
         Transaction transaction = null;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
-            Trainee trainee = session.createQuery( "SELECT t FROM Trainee t WHERE t.userName = :userName", Trainee.class)
-                    .setParameter("userName", userName)
-                    .uniqueResult();
+            Trainee trainee = this.getTraineeByUserName(userName, password);
 
-            if (trainee != null) {
+            if (trainee.getUserName().equals(userName) && trainee.getPassword().equals(password)) {
                 Boolean isActive = !trainee.getIsActive(); // Toggle status
 
                 // Update trainee's active status
@@ -97,7 +129,7 @@ public class TraineeRepositoryImpl implements TraineeRepository {
                 transaction.commit();
                 System.out.println("Trainee status updated successfully.");
             } else {
-                System.out.println("Trainee not found with username: " + userName);
+                logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
             }
         } catch (Exception e) {
             if (transaction != null) {
@@ -109,41 +141,98 @@ public class TraineeRepositoryImpl implements TraineeRepository {
     }
 
     @Override
-    public List<Training> getTraineeTrainings(String userName, String fromDate, String toDate, String trainerUserName) {
-        return null;
+    public List<Training> getTraineeTrainingList(String traineeName, String password, String fromDate, String toDate, String trainerName) {
+        try (Session session = sessionFactory.openSession()) {
+            //  Check userName and password matching
+            Trainee trainee = this.getTraineeByUserName(traineeName, password);
+            if (trainee.getUserName().equals(traineeName) && trainee.getPassword().equals(password)) {
+                logger.log(Level.INFO, "Successfull authentification for trainer: " + trainee.getUserName());
+                return session.createQuery(
+                                "SELECT t FROM Training t " +
+                                        "WHERE t.trainee.userName = :traineeName " +
+                                        "AND t.trainingDate >= :fromDate " +
+                                        "AND t.trainingDate <= :toDate " +
+                                        "AND t.trainer.userName = :trainerName", Training.class)
+                        .setParameter("traineeName", traineeName)
+                        .setParameter("fromDate", LocalDate.parse(fromDate))
+                        .setParameter("toDate", LocalDate.parse(toDate))
+                        .setParameter("trainerName", trainerName)
+                        .getResultList();
+            } else {
+                logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
+                return null;
+            }
+        }
     }
+
 
     @Override
     public Trainee updateTrainee(Trainee trainee) {
-        Trainee existingTrainee = getTraineeByUserName(trainee.getUserName());
+        Trainee existingTrainee = getTraineeByUserName(trainee.getUserName(), trainee.getPassword());
+        //  Check userName and password matching
+        if (existingTrainee.getUserName().equals(trainee.getUserName()) && existingTrainee.getPassword().equals(trainee.getPassword())) {
+            logger.log(Level.INFO, "Successfull authentification for trainer: " + trainee.getUserName());
+            existingTrainee.setFirstName(trainee.getFirstName());
+            existingTrainee.setLastName(trainee.getLastName());
+            existingTrainee.setPassword(trainee.getPassword());
+            existingTrainee.setIsActive(trainee.getIsActive());
+            existingTrainee.setDateOfBirth(trainee.getDateOfBirth());
+            existingTrainee.setAddress(trainee.getAddress());
 
-        existingTrainee.setFirstName(trainee.getFirstName());
-        existingTrainee.setLastName(trainee.getLastName());
-        existingTrainee.setPassword(trainee.getPassword());
-        existingTrainee.setIsActive(trainee.getIsActive());
-        existingTrainee.setDateOfBirth(trainee.getDateOfBirth());
-        existingTrainee.setAddress(trainee.getAddress());
-
-        try (Session session = sessionFactory.openSession()) {
-            Transaction transaction = session.beginTransaction();
-            Trainee updatedTrainee = (Trainee) session.merge(existingTrainee);
-            transaction.commit();
-            return updatedTrainee;
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw e;
+            try (Session session = sessionFactory.openSession()) {
+                Transaction transaction = session.beginTransaction();
+                Trainee updatedTrainee = (Trainee) session.merge(existingTrainee);
+                transaction.commit();
+                return updatedTrainee;
+            } catch (Exception e) {
+                e.printStackTrace();
+                throw e;
+            }
+        } else {
+            logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
+            return null;
         }
     }
 
     @Override
-    public List<Trainee> findAll() {
-        return null;
+    public List<Trainer> getNotAssignedTrainers(String traineeUserName, String password) {
+        try (Session session = sessionFactory.openSession()) {
+            //  Check userName and password matching
+            Trainee trainee = this.getTraineeByUserName(traineeUserName, password);
+            if (trainee.getUserName().equals(traineeUserName) && trainee.getPassword().equals(password)) {
+                logger.log(Level.INFO, "Successfull authentification for trainer: " + trainee.getUserName());
+                return session.createQuery(
+                                "SELECT tr FROM Trainer tr " +
+                                        "WHERE tr.id NOT IN (" +
+                                        "   SELECT ttr.id FROM Trainee t " +
+                                        "   JOIN t.trainers ttr " +
+                                        "   WHERE t.user.userName = :traineeUserName" +
+                                        ")",
+                                Trainer.class)
+                        .setParameter("traineeUserName", traineeUserName)
+                        .getResultList();
+            } else {
+                logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
+                return null;
+            }
+        }
     }
 
-
-
     @Override
-    public void delete(Long id) {
+    public void updateTrainersList(String traineeUserName, String password, List<Trainer> trainers) {
+        try (Session session = sessionFactory.openSession()) {
+            Transaction transaction = session.beginTransaction();
+            //  Check userName and password matching
+            Trainee trainee = this.getTraineeByUserName(traineeUserName, password);
+            if (trainee.getUserName().equals(traineeUserName) && trainee.getPassword().equals(password)) {
+                logger.log(Level.INFO, "Successfull authentification for trainer: " + trainee.getUserName());
 
+                trainee.setTrainers(trainers);
+                session.update(trainee);
+            } else {
+                logger.log(Level.INFO, "Incorrect userName and password for trainee: " + trainee.getUserName());
+            }
+            transaction.commit();
+        }
     }
 }
